requirements.txt
numpy==1.26.4
pandas==2.2.3


# frontend
streamlit==1.39.0
## for nice map tiles
folium==0.18.0
streamlit_folium==0.23.1

# backend 
datasets==3.0.2
## FSM
transitions==0.9.2 

# data manipulation 
pandas==2.2.3

# running ML models

## to use ML models hosted on HF 
huggingface-hub==0.26.1
transformers==4.46.0
#torch
## +minimal reqs for locally running model - with PIL load
pytorch-lightning<=1.5.10
timm==0.5.4
pillow==10.4.0

## also need opencv for handling images in the later model with preprocessing
opencv-python-headless==4.5.5.64
albumentations==1.1.0

# for env variables
python-dotenv==1.1.0

# documentation: mkdocs
mkdocs~=1.6.0
mkdocstrings[python]>=0.25.1
mkdocs-material~=9.5.27
mkdocs-homepage-copier~=1.0.0
